backend/requirements.txt
blinker==1.7.0
CacheControl==0.14.0
cachetools==5.5.0
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.3.2
click==8.1.7
cryptography==43.0.1
firebase-admin==6.5.0
Flask==3.0.2
Flask-Cors==5.0.0
google-api-core==2.19.2
google-api-python-client==2.145.0
google-auth==2.34.0
google-auth-httplib2==0.2.0
google-cloud-core==2.4.1
google-cloud-firestore==2.18.0
google-cloud-storage==2.18.2
google-crc32c==1.6.0
google-resumable-media==2.7.2
googleapis-common-protos==1.65.0
grpcio==1.66.1
grpcio-status==1.66.1
gunicorn==20.1.0
httplib2==0.22.0
idna==3.8
itsdangerous==2.1.2
Jinja2==3.1.3
MarkupSafe==2.1.5
msgpack==1.1.0
proto-plus==1.24.0
protobuf==5.28.1
pyasn1==0.6.1
pyasn1_modules==0.4.1
pycparser==2.22
PyJWT==2.9.0
pyparsing==3.1.4
python-dotenv==1.0.1
requests==2.32.3
rsa==4.9
uritemplate==4.1.1
urllib3==2.2.3
Werkzeug==3.0.1
